# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smaki CRM API Documentation",
        "version": "0.3"
    },
    "servers": [
        {
            "url": "https://api.smaki.com.ua/api"
        }
    ],
    "paths": {
...
        "/order-items": {
            "get": {
                "tags": [
                    "Order.Items"
                ],
                "summary": "Список позиций из заказов для приготовления",
                "description": "Доступна поварам. Выдает позиции для приготовления в зависимости от типа товара и кухни повара",
                "operationId": "902a3de3f969553957a25940111279d2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID",
                                            "type": "integer"
                                        },
                                        "order_id": {
                                            "description": "ID заказа",
                                            "type": "integer"
                                        },
                                        "product_id": {
                                            "description": "ID товара",
                                            "type": "integer"
                                        },
                                        "quantity": {
                                            "description": "Количество",
                                            "type": "integer"
                                        },
                                        "sum": {
                                            "description": "Стоимость",
                                            "type": "integer"
                                        },
                                        "comment": {
                                            "description": "Комментарий",
                                            "type": "string"
                                        },
                                        "cook_id": {
                                            "description": "ID повара",
                                            "type": "integer"
                                        },
                                        "status": {
                                            "description": "Статус",
                                            "type": "string"
                                        },
                                        "product": {
                                            "description": "Объект товара",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": {
                                            "0": {
                                                "id": 28,
                                                "order_id": 21,
                                                "product_id": "88526c81-9dbd-4d8b-91df-6b12cebf5c0a",
                                                "quantity": 2,
                                                "sum": 150,
                                                "comment": null,
                                                "created_at": "2021-10-25T20:32:16.000000Z",
                                                "updated_at": "2021-10-25T20:32:16.000000Z",
                                                "cook_id": null,
                                                "status": "new",
                                                "product": {
                                                    "id": "88526c81-9dbd-4d8b-91df-6b12cebf5c0a",
                                                    "restaurant": "smaki",
                                                    "article": "art-2",
                                                    "title_ua": "Саке макі",
                                                    "title_ru": "Саке макі",
                                                    "weight": 250,
                                                    "weight_type": "г",
                                                    "category_sync_id": null,
                                                    "type_sync_id": "sushi",
                                                    "description_ua": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse blandit hendrerit mi eget aliquam. Mauris volutpat sem augue, quis tincidunt leo vehicula ut. Pellentesque convallis vel eros vitae luctus. Curabitur commodo fringilla risus, consectetur varius enim porttitor vel. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Phasellus magna sem, efficitur quis interdum quis, viverra in purus. Sed finibus odio euismod felis gravida mollis. Praesent ante sapien, ornare sed consectetur at, euismod eget nulla. Suspendisse accumsan imperdiet sagittis. Maecenas quis mi tortor. Integer vel ante ut nisi pulvinar tincidunt id sed metus. Nunc non erat ut tortor finibus gravida quis quis magna. Suspendisse varius odio sed leo scelerisque pretium. Praesent auctor odio sit amet leo posuere, nec tempus risus tincidunt.",
                                                    "description_ru": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse blandit hendrerit mi eget aliquam. Mauris volutpat sem augue, quis tincidunt leo vehicula ut. Pellentesque convallis vel eros vitae luctus. Curabitur commodo fringilla risus, consectetur varius enim porttitor vel. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Phasellus magna sem, efficitur quis interdum quis, viverra in purus. Sed finibus odio euismod felis gravida mollis. Praesent ante sapien, ornare sed consectetur at, euismod eget nulla. Suspendisse accumsan imperdiet sagittis. Maecenas quis mi tortor. Integer vel ante ut nisi pulvinar tincidunt id sed metus. Nunc non erat ut tortor finibus gravida quis quis magna. Suspendisse varius odio sed leo scelerisque pretium. Praesent auctor odio sit amet leo posuere, nec tempus risus tincidunt.",
                                                    "created_at": "2021-10-01T16:25:57.000000Z",
                                                    "updated_at": "2021-10-01T16:25:57.000000Z",
                                                    "image": "http://smaki.local/images/default.jpg",
                                                    "media": []
                                                }
                                            },
                                            "links": {
                                                "first": "https://api.smaki.com.ua/api/orders?page=1",
                                                "last": "https://api.smaki.com.ua/api/orders?page=1",
                                                "prev": null,
                                                "next": null
                                            },
                                            "meta": {
                                                "current_page": 1,
                                                "from": 1,
                                                "last_page": 1,
                                                "links": [],
                                                "path": "https://api.smaki.com.ua/api/orders",
                                                "per_page": 20,
                                                "to": 11,
                                                "total": 11
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/order-items-statuses": {
            "get": {
                "tags": [
                    "Order.Items"
                ],
                "summary": "Статусы позиций в заказе",
                "operationId": "63cb5685259889dfaf52584552de5802",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "Статус",
                                            "type": "string"
                                        },
                                        "title": {
                                            "description": "Название статуса",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": [
                                            {
                                                "name": "new",
                                                "title": "Новый заказ"
                                            },
                                            {
                                                "name": "in_process",
                                                "title": "Готовится"
                                            },
                                            {
                                                "name": "ready",
                                                "title": "Приготовлен"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}